Moving some parsing from Sale to Invoice
