+++ src/receipt/Invoice.js
@@ -1,3 +1,6 @@
+import Item from "./Item.js"
+import Receipt from "./Receipt.js"
+import Share from "./Share.js"
 import Sale from "./Sale.js"
 import NodesGenerator from "./xml/NodesGenerator.js"
 
@@ -222,6 +225,95 @@ class Invoice extends Sale {
 		NodesGenerator.generateTotal(this)
 
 		NodesGenerator.generateLines(this)
+	}
+
+	/**
+	 * @return xmlDoc parsed
+	 */
+	fromXml(xmlContent) {
+		// All about sale
+		const xmlDoc = super.fromXml(xmlContent)
+
+		{
+			const orderReference = xmlDoc.getElementsByTagNameNS(Receipt.namespaces.cac, "OrderReference")[0];
+			if (orderReference) {
+				const orderReferenceId = orderReference.getElementsByTagNameNS(Receipt.namespaces.cbc, "ID")[0]?.textContent;
+
+				if (orderReferenceId) {
+					this.setOrderReference(orderReferenceId)
+					const orderReferenceText = orderReference.getElementsByTagNameNS(Receipt.namespaces.cbc, "CustomerReference")[0]?.textContent;
+					if (orderReferenceText) {
+						this.setOrderReferenceText(orderReferenceText)
+					}
+				}
+			}
+		}
+
+		// All items. Always present.
+		const items = xmlDoc.getElementsByTagNameNS(Receipt.namespaces.cac, "InvoiceLine")
+		for (let i = 0; i < items.length; i++) {
+			const item = new Item( items[i].getElementsByTagNameNS(Receipt.namespaces.cbc, "Description")[0]?.textContent || "" )
+			item.setQuantity( items[i].getElementsByTagNameNS(Receipt.namespaces.cbc, "InvoicedQuantity")[0]?.textContent || "" )
+			item.setUnitValue(
+				items[i].getElementsByTagNameNS(Receipt.namespaces.cac, "Price")[0]?.getElementsByTagNameNS(Receipt.namespaces.cbc, "PriceAmount")[0]?.textContent,
+				false
+			)
+			item.setUnitCode( items[i].getElementsByTagNameNS(Receipt.namespaces.cbc, "InvoicedQuantity")[0]?.getAttribute("unitCode") || "" )
+
+			// Warning because there are many tags with same name
+			item.setIgvPercentage( parseInt(items[i].getElementsByTagNameNS(Receipt.namespaces.cbc, "Percent")[0]?.textContent) )
+			item.setExemptionReasonCode( parseInt(items[i].getElementsByTagNameNS(Receipt.namespaces.cbc, "TaxExemptionReasonCode")[0]?.textContent) )
+
+			item.calcMounts()
+			this.addItem(item)
+		}
+
+		{ // about detractions
+			// possible deduction
+			const paymentMean = xmlDoc.getElementsByTagNameNS(Receipt.namespaces.cac, "PaymentMeans")[0]
+			if (paymentMean) { // exists deduction
+				const id = paymentMean.getElementsByTagNameNS(Receipt.namespaces.cbc, "ID")[0]?.textContent
+				if (id == "Detraccion") { // deduction exists
+					// look for more about this deduction
+					const paymentTerms = xmlDoc.getElementsByTagNameNS(Receipt.namespaces.cac, "PaymentTerms")
+					for (const paymentTerm of paymentTerms) {
+						if (paymentTerm.getElementsByTagNameNS(Receipt.namespaces.cbc, "ID")[0].textContent != "Detraccion") {
+							continue
+						}
+
+						// we found it
+						this.setDetractionPercentage(parseInt(paymentTerm.getElementsByTagNameNS(Receipt.namespaces.cbc, "PaymentPercent")[0].textContent))
+						this.calcDetractionAmount()
+						break // then nothing else
+					}
+				}
+			}
+		}
+
+		{ // check if there are shares
+			const paymentTerms = xmlDoc.getElementsByTagNameNS(Receipt.namespaces.cac, "PaymentTerms") // always exists
+			for (let i = 0; i < paymentTerms.length; ++i) {
+				if (paymentTerms[i].getElementsByTagNameNS(Receipt.namespaces.cbc, "ID")[0].textContent == "FormaPago") {
+					// If there is Credito means that next siblings are amounts
+					if (paymentTerms[i].getElementsByTagNameNS(Receipt.namespaces.cbc, "PaymentMeansID")[0].textContent == "Credito") {
+						++i; // set index to next sibling
+						// iterate posible remaining shares
+						for (; i < paymentTerms.length; ++i) {
+							// if FormaPago is not found, means we don't have any share
+							if (paymentTerms[i].getElementsByTagNameNS(Receipt.namespaces.cbc, "ID")[0].textContent != "FormaPago") {
+								break
+							}
+							const share = new Share()
+							// capture date and amount
+							share.setAmount(paymentTerms[i].getElementsByTagNameNS(Receipt.namespaces.cbc, "Amount")[0].textContent)
+							const dateParts = paymentTerms[i].getElementsByTagNameNS(Receipt.namespaces.cbc, "PaymentDueDate")[0].textContent.split('-') // split in year, month and day
+							share.setDueDate(new Date(dateParts[0], dateParts[1] - 1, dateParts[2]))
+							this.addShare(share)
+						}
+					}
+				}
+			}
+		}
 	}
 }
